First commit; add Vagrantfile for VM configs.
This first commit is my first attempt at a Vagrantfile.  It expects VirtualBox and Debian Jessie 64-bit.  For speed and efficiency, git only clones in the VM with `--depth=1` and does not pull full history.

The command structure is broken down into `$prereqsteps`, then `$compsteps` for compilation, and the final `$ldcacherebuild` so the necessary .so libraries are load.  This is done just to increase reaability.